+++ tests/Support/UnionOptionalDependency.php
@@ -7,7 +7,7 @@ namespace Yiisoft\Definitions\Tests\Support;
 final class UnionOptionalDependency
 {
     public function __construct(
-        private $value = null
+        private string|ColorInterface $value = 'test'
     ) {
     }