+++ tests/Unit/ParameterDefinitionTest.php
@@ -16,6 +16,7 @@ use Yiisoft\Definitions\Exception\InvalidConfigException;
 use Yiisoft\Definitions\Exception\NotInstantiableException;
 use Yiisoft\Definitions\ParameterDefinition;
 use Yiisoft\Definitions\Tests\Support\CircularReferenceExceptionDependency;
+use Yiisoft\Definitions\Tests\Support\ColorInterface;
 use Yiisoft\Definitions\Tests\Support\GearBox;
 use Yiisoft\Definitions\Tests\Support\RuntimeExceptionDependency;
 use Yiisoft\Definitions\Tests\Support\Car;
@@ -334,11 +335,21 @@ final class ParameterDefinitionTest extends TestCase
     public function testResolveOptionalUnionType(): void
     {
         $definition = new ParameterDefinition(
+            $this->getFirstParameter(static fn (string|ColorInterface $value = 'test') => true)
+        );
+        $container = new SimpleContainer();
+
+        $this->assertSame('test', $definition->resolve($container));
+    }
+
+    public function testResolveOptionalPromotedPropertyUnionType(): void
+    {
+        $definition = new ParameterDefinition(
             $this->getFirstConstructorParameter(UnionOptionalDependency::class)
         );
         $container = new SimpleContainer();
 
-        $this->assertNull($definition->resolve($container));
+        $this->assertSame('test', $definition->resolve($container));
     }
 
     public function testResolveUnionBuiltin(): void